job_worker after deployment
