--- job_worker/README.md
@@ -1,1 +1,0 @@
-# Job worker
