+++ job_worker/compile/constant.py
@@ -1,0 +1,7 @@
+GB = 1024*1024*1024
+
+PROJECT_ID = "swe-team9"
+COMPILE_SUBSCRIPTION_ID = "compileTopic-sub"
+MEASURE_SUBSCRIPTION_ID = "measureTopic-sub"
+FIREBASE_DB_URL = 'https://swe-team9-default-rtdb.firebaseio.com'
+STORAGE_BUCKET_NAME = 'earth-saver'
