--- job_worker/constant.py
@@ -1,1 +1,0 @@
-GB = 1024*1024*1024
